Deduplicate some methods
