+++ rustfmt.toml
@@ -1,0 +1,3 @@
+# Imports
+imports_granularity = "Crate"
+reorder_imports = true
